[Chore] #37 - 자잘한 수정사항 반영
