+++ Gajeonggapsseo-iOS/Extensions/Date+Extension.swift
@@ -10,7 +10,7 @@ import Foundation
 extension Date {
     func toYearMonthDayString() -> String {
         let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
+        dateFormatter.dateFormat = "yyyy-MM-dd HH:mm"
         dateFormatter.locale = Locale(identifier: "ko_KR")
         dateFormatter.timeZone = TimeZone(identifier: "Asia/Seoul")
         return dateFormatter.string(from: self)
@@ -22,5 +22,22 @@ extension Date {
         dateFormatter.locale = Locale(identifier: "ko_KR")
         dateFormatter.timeZone = TimeZone(identifier: "Asia/Seoul")
         return dateFormatter.date(from: dateString)
+    }
+    
+    static func fromOperationTimeToDateString(_ dateString: String) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "HH시"
+        dateFormatter.locale = Locale(identifier: "ko_KR")
+        dateFormatter.timeZone = TimeZone(identifier: "Asia/Seoul")
+        let opertationTimeDate = dateFormatter.date(from: dateString)
+        dateFormatter.dateFormat = "HH:00"
+        return dateFormatter.string(from: opertationTimeDate ?? Date())
+    }
+    
+    static func fromOperatingHoursToDateString(_ dateString: String) -> String {
+        var formattedString = dateString
+        formattedString = formattedString.replacingOccurrences(of: "시", with: ":00")
+        formattedString = formattedString.replacingOccurrences(of: "~", with: " ~ ")
+        return formattedString
     }
 }
