+++ Gajeonggapsseo-iOS/Views/Map/Sheet/JejuRecycleSheetView.swift
@@ -43,7 +43,7 @@ struct JejuRecycleSheetView: View {
                 }
                 .padding(.trailing, 22)
                 VStack(alignment: .leading, spacing: 9) {
-                    Text("\(center.operatingHours)").font(.system(size: 18, weight: .regular))
+                    Text("\(Date.fromOperatingHoursToDateString(center.operatingHours))").font(.system(size: 18, weight: .regular))
                     Text("\(center.address.replacingOccurrences(of: "제주특별자치도 ", with: ""))").font(.system(size: 18, weight: .regular))
                     Spacer()
                 }
